add test for maximum_projection_from_path
